# backend/docs/swagger.yaml
basePath: /api
definitions:
  rest.Err:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/rest.ErrDetail'
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
  rest.ErrDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  rest.ErrResponse:
    properties:
      error:
        $ref: '#/definitions/rest.Err'
    type: object
...
  rest.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  rest.TaskTagAssignmentRequest:
    properties:
      tag_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Task scheduling API server.
  title: task2todaytodo API
  version: "1.0"
paths:
...
  /tasks/{id}/tags:add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.TaskTagAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Add tag to task
      tags:
      - Tasks
  /tasks/{id}/tags:remove:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.TaskTagAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Remove tag from task
      tags:
      - Tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"